Adding SSL

+++ src/main/java/JSocket/Common/IO/ConnectionIO.java
@@ -164,7 +164,7 @@ public class ConnectionIO implements Closeable,AutoCloseable,Readable {
     private int[] toIntArray(byte[] array) {
         int[] arr = new int[array.length];
         for (int i=0;i<array.length;i++){
-            arr[i] = (int)array[i];
+            arr[i] = array[i];
         }
         return arr;
     }
