+++ src/main/java/JSocket/Common/IO/DataFrameOutputStream.java
@@ -24,7 +24,7 @@ public class DataFrameOutputStream extends OutputStream {
     private int[] toIntArray(byte[] array) {
         int[] arr = new int[array.length];
         for (int i=0;i<array.length;i++){
-            arr[i] = (int)array[i];
+            arr[i] = array[i];
         }
         return arr;
     }
